Remove LedState
